% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASE HMM.R
\name{ASE.HMM}
\alias{ASE.HMM}
\title{Dection of Differentially Allelic Expressed Regions (DAERs) by hmmASE}
\usage{
ASE.HMM(norm.M, norm.P, abnorm.M, abnorm.P, ex.rm = FALSE, cutoff,
  decoding = "Global", rep = c(4, 4), min.length = 0, min.SNP = 1,
  max.dist = 1e+20)
}
\arguments{
\item{norm.M}{A \code{dataframe} contains the read counts for maternal allele in normal group. The dimension of this \code{dataframe} can vary with the number of available biological replicates but the first 4 columns should be Chromosome, GeneID, Gene_name, and Position.}

\item{norm.P}{A \code{dataframe} contains the read counts for paternal allele in normal group, with similar structure with \code{norm.M}.}

\item{abnorm.M}{A \code{dataframe} contains the read counts for maternal allele in abnormal group, with similar structure with \code{norm.M} and \code{norm.P}.}

\item{abnorm.P}{A \code{dataframe} contains the read counts for paternal allele in abnormal group, with similar structure with other three input datasets.}

\item{ex.rm}{By setting \code{ex.rm=TRUE}, the outliers would be removed from the hidden markov model in the EM steps, and the status prediction would be directly carried out based on the sign of observations. Default value is \code{ex.rm=FALSE} which means the observations from all the SNPs would used in the prediction of the best sequence of hidden status.}

\item{cutoff}{If \code{ex.rm=TRUE}, only the observations of SNPs which have \eqn{|O_1|<= cutoff} would be included in the hidden markov model. Setting \code{cutoff=7} or larger is equivalent to setting \code{ex.rm=FALSE}.}

\item{decoding}{Specification of the decoding method for Hidden markov model, which can be either "Local" or "Global", with the latter one referring to the Viterbi algorithm. The default is "Global".}

\item{rep}{A vector indicates the number of biological replicates for normal and abnormal groups with default value being (4,4).}

\item{min.length}{The minimum length of a differentially allelic expressed region, defaultly set to be 0.}

\item{min.SNP}{The minimum number of SNPs in a differentially allelic expressed region with a default value of 1.}

\item{max.dist}{The maximum gap of two consecutive SNPs in a differentially allelic expressed region, and the default value \eqn{10^{20}}.}
}
\value{
A \code{list} contains the following elements:
\describe{
\item{\code{region}}{A \code{dataframe} contains 6 columns, i.e., region.cnt, region.start, region.end, region.state,num.SNP and length, which means the index of regions, the start and end position of a region, the predicted state of a region, the number of SNPs and the length in bps of a region, respectively.}
\item{\code{DAER.res}}{A \code{dataframe} summarize the input data and predicted hidden status.}
}
}
\description{
A function to implement the hmmASE method.
}
