% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estep.R
\name{Estep}
\alias{Estep}
\title{The E-step in the EM algorithm}
\usage{
Estep(obs, par)
}
\arguments{
\item{obs}{A S by 5 \code{dataframe} returned by \code{data.trans} function where S is total number of SNPs.}

\item{par}{A \code{list} contains the current estimates of parameters.}
}
\value{
A \code{list} contains the following elements:
\describe{
\item{\code{e.loglik}}{The expectation of log-likelihood.}
\item{\code{P.ks}}{a S by 9 matrix contains \eqn{P(S_s=k|O,\theta)}}
\item{\code{P.kls}}{a S by 81 matrix contains \eqn{P(S_s=k,S_(s+1)=l|O,\theta)} }
}
}
\description{
A function to compute the expectation of the logarithm of complete likelihood w.r.t the conditional distribution of hidden status given the observed data and parameters.
}
