% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep.R
\name{Mstep}
\alias{Mstep}
\title{The M-step in the EM algorithm}
\usage{
Mstep(obs, par.current, e.res)
}
\arguments{
\item{obs}{A S by 5 \code{dataframe} returned by \code{data.trans} function where S is total number of SNPs.}

\item{par.current}{A \code{list} contains the current estimates of parameters.}

\item{e.res}{A \code{list} contains the current results of E-step given by function \code{Estep()}.}
}
\value{
A \code{list} contains the updated values of parameters.
}
\description{
A function to update the parameters.
}
